+++ helicase/src/parser/fastq.rs
@@ -710,8 +710,7 @@ mod tests {
         .ignore_headers()
         .dna_string()
         .split_non_actg()
-        .config()
-        & !RETURN_RECORD;
+        .config();
     const CONFIG_STRING_ACTG_MERGE: Config = ParserOptions::default()
         .ignore_headers()
         .dna_string()
@@ -720,8 +719,7 @@ mod tests {
     const CONFIG_COLUMNAR: Config = ParserOptions::default()
         .ignore_headers()
         .dna_columnar()
-        .config()
-        & !RETURN_RECORD;
+        .config();
     const CONFIG_COLUMNAR_MERGE: Config = ParserOptions::default()
         .ignore_headers()
         .dna_columnar()
@@ -730,8 +728,7 @@ mod tests {
     const CONFIG_PACKED: Config = ParserOptions::default()
         .ignore_headers()
         .dna_packed()
-        .config()
-        & !RETURN_RECORD;
+        .config();
     const CONFIG_PACKED_MERGE: Config = ParserOptions::default()
         .ignore_headers()
         .dna_packed()